+++ not-supported/not-supported_menu.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# T&M Hansson IT AB © - 2020, https://www.hanssonit.se/
+# T&M Hansson IT AB © - 2021, https://www.hanssonit.se/
 
 true
 SCRIPT_NAME="Not-supported Menu"
@@ -26,7 +26,6 @@ So please run them on your own risk. Feedback is more than welcome, though and c
 Choose which one you want to execute.
 $CHECKLIST_GUIDE" "$WT_HEIGHT" "$WT_WIDTH" 4 \
 "Bitlocker Mount" "(Mount Bitlocker encrypted drives)" OFF \
-"ClamAV" "(Antivirus for Nextcloud and files)" OFF \
 "NTFS Format" "(Format drives to NTFS)" OFF \
 "NTFS Mount" "(Mount NTFS drives)" OFF \
 "Backup Viewer" "(View your Backups)" OFF \
@@ -44,10 +43,6 @@ case "$choice" in
     *"Bitlocker Mount"*)
         print_text_in_color "$ICyan" "Downloading the Bitlocker Mount script..."
         run_script NOT_SUPPORTED_FOLDER bitlocker-mount
-    ;;&
-    *"ClamAV"*)
-        print_text_in_color "$ICyan" "Downloading the ClamAV script..."
-        run_script APP clamav
     ;;&
     *"NTFS Format"*)
         print_text_in_color "$ICyan" "Downloading the NTFS Format script..."