+++ not-supported/pi-hole.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-# T&M Hansson IT AB © - 2020, https://www.hanssonit.se/
-# Copyright © 2020 Simon Lindner (https://github.com/szaimen)
+# T&M Hansson IT AB © - 2021, https://www.hanssonit.se/
+# Copyright © 2021 Simon Lindner (https://github.com/szaimen)
 
 # shellcheck disable=2016,2034,2059,2178
 true
@@ -230,6 +230,7 @@ mapfile -t INSTALLED <<< "${INSTALLED[@]}"
 cat << PIHOLE_UPDATE >> "$SCRIPTS/update.sh"
 
 #Pi-hole-start - Please don't remove or change this line
+PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin
 check_command pihole -up
 check_command sed -i 's|^server\.port.*|server\.port = 8093|' /etc/lighttpd/lighttpd.conf
 sleep 5 # Wait for lighttpd